+++ cli/cmd/error_log.go
@@ -11,7 +11,6 @@ import (
 
 // Logs an error message in client database, to be accessible for viewing by user later
 func LogError(db *database.Queries, cmd *cobra.Command, err error, userMsg string) {
-
 	fmt.Println(userMsg)
 
 	timestamp := time.Now().UTC().Truncate(time.Second)