+++ docs/changelog.md
@@ -7,10 +7,12 @@ All notable changes to this project will be documented in this file.
 - CLI
     - Log command for in-depth viewing of error logs from database
     - Auto updating/syncing of user items upon user login
+    - Commands for setting/clearing default item/account values, that may be used in place of command arguments
 
 ### Changed
 - CLI
     - User viewable errors are more generic, with verbose error details stored in local database
+    - Certain command arguments can now be ommitted, if default values have been set in settings
 
 ### Fixed
 - CLI